replace thorn by th
Now it's replaced to a `b` (because of the shape?). Although this letter doesn't occur in Belgian railway Stations (and Iceland doesn't have trains anyway) this doesn't really matter a lot. But [thorn](https://en.wikipedia.org/wiki/Thorn_(letter)) is usually transformed to ascii as `th`, and not `b` or `p`.